feat: print event buttons by path
