+++ client/src/components/atom/main/FavoriteButton.jsx
@@ -1,11 +1,43 @@
-import React from 'react';
+import React, { useState, useRef, useEffect, useCallback } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { welcomeModal } from '../../../redux/modules/modal';
 import { IconButton } from '@material-ui/core';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 
-export default function FavoriteButton() {
+export default function FavoriteButton({ event }) {
+  const [select, setSelect] = useState(false);
+  const user = useSelector((state) => state.auth.user);
+  const dispatch = useDispatch();
+  const buttonRef = useRef();
+  const eventId = event && event._id;
+  const userId = user && user._id;
+  const userIds =
+    event && event.liked_users && event.liked_users.map((user) => user._id);
+
+  useEffect(() => {
+    event &&
+      event.liked_users.map((user) => {
+        user._id === userId && setSelect(true);
+      });
+  }, [userId]);
+
+  const viewModal = useCallback(() => {
+    dispatch(welcomeModal('이 기능은 회원만 가능해요 😉'));
+  }, [dispatch]);
+
+  const clickFav = () => {
+    if (user === null) {
+      viewModal();
+      return;
+    }
+    setSelect(!select);
+  };
+
   return (
-    <IconButton aria-label='favorite'>
-      <FavoriteIcon />
-    </IconButton>
+    <div className={select ? 'act favoriteButton-wrap' : 'favoriteButton-wrap'}>
+      <IconButton aria-label="favorite" onClick={clickFav} ref={buttonRef}>
+        <FavoriteIcon />
+      </IconButton>
+    </div>
   );
 }
