+++ client/src/components/atom/main/SettingButtons.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import { IconButton } from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import PeopleIcon from '@material-ui/icons/People';
+import EditIcon from '@material-ui/icons/Edit';
+import DeleteIcon from '@material-ui/icons/Delete';
+
+export default function SettingButtons() {
+  return (
+    <>
+      <IconButton aria-label="enlistedUsers">
+        <PeopleIcon />
+      </IconButton>
+      <IconButton aria-label="edit">
+        <EditIcon />
+      </IconButton>
+      <IconButton aria-label="delete">
+        <DeleteIcon />
+      </IconButton>
+    </>
+  );
+}
